// web/src/api/generated/public/models/FlashSaleActivityResponse.ts
export interface FlashSaleActivityResponse {
    /**
     *
     * @type {string}
     * @memberof FlashSaleActivityResponse
     */
    activityCode: string;
    /**
     *
     * @type {string}
     * @memberof FlashSaleActivityResponse
     */
    activityId: string;
    /**
     *
     * @type {number}
     * @memberof FlashSaleActivityResponse
     */
    availableStock: number;
    /**
     *
     * @type {string}
     * @memberof FlashSaleActivityResponse
     */
    category: string | null;
    /**
     *
     * @type {string}
     * @memberof FlashSaleActivityResponse
     */
    description: string | null;
    /**
     *
     * @type {Date}
     * @memberof FlashSaleActivityResponse
     */
    endsAt: Date;
    /**
     *
     * @type {number}
     * @memberof FlashSaleActivityResponse
     */
    perUserLimit: number;
    /**
     *
     * @type {string}
     * @memberof FlashSaleActivityResponse
     */
    phase: FlashSaleActivityResponsePhaseEnum;
    /**
     *
     * @type {string}
     * @memberof FlashSaleActivityResponse
     */
    productId: string;
    /**
     *
     * @type {string}
     * @memberof FlashSaleActivityResponse
     */
    productName: string;
    /**
     *
     * @type {string}
     * @memberof FlashSaleActivityResponse
     */
    salePrice: string;
    /**
     *
     * @type {Date}
     * @memberof FlashSaleActivityResponse
     */
    serverTime: Date;
    /**
     *
     * @type {Date}
     * @memberof FlashSaleActivityResponse
     */
    startsAt: Date;
    /**
     *
     * @type {string}
     * @memberof FlashSaleActivityResponse
     */
    status: string;
}


/**
 * @export
 */
export const FlashSaleActivityResponsePhaseEnum = {
    Upcoming: 'UPCOMING',
    Live: 'LIVE',
    SoldOut: 'SOLD_OUT',
    Expired: 'EXPIRED'
} as const;
export type FlashSaleActivityResponsePhaseEnum = typeof FlashSaleActivityResponsePhaseEnum[keyof typeof FlashSaleActivityResponsePhaseEnum];